student/requirements.txt
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0